refactor sendPhoneCodeUnion to setPhoneNumberUnion
