+++ api-js/src/user/unions/index.js
@@ -1,6 +1,6 @@
 export * from './authenticate-union'
 export * from './reset-password-union'
-export * from './send-phone-code-union'
+export * from './set-phone-number-union'
 export * from './sign-in-union'
 export * from './sign-up-union'
 export * from './update-user-password-union'
